Move `Nil` to `fitness` module
