+++ packages/brace-ec/src/core/mod.rs
@@ -1,3 +1,4 @@
+pub mod fitness;
 pub mod generation;
 pub mod individual;
 pub mod operator;
